+++ Import/Manager/Konnekting.cs
@@ -166,7 +166,7 @@ namespace Kaenx.DataContext.Import.Manager
                     model.ManufacturerId = manuId;
                     context.Devices.Add(model);
 
-                    Dictionary<int, IValues> Id2Values = new Dictionary<int, IValues>();
+                    Dictionary<long, IValues> Id2Values = new Dictionary<long, IValues>();
                     List<IDynChannel> Channels = new List<IDynChannel>();
                     ChannelIndependentBlock channel = new ChannelIndependentBlock() {  IsVisible = true, HasAccess = true };
                     Channels.Add(channel);
@@ -523,7 +523,7 @@ namespace Kaenx.DataContext.Import.Manager
             return paraType;
         }
 
-        private void CalcDefaultVisibility(ChannelIndependentBlock channel, Dictionary<int, IValues> values){
+        private void CalcDefaultVisibility(ChannelIndependentBlock channel, Dictionary<long, IValues> values){
             foreach(ParameterBlock block in channel.Blocks) {
                 block.IsVisible = FunctionHelper.CheckConditions(block.Conditions, values);
